Implement ClickHouse and DuckDB interaction repositories
- Introduced `ClickHouseInteractionRepository` and `DuckDbInteractionRepository` to handle interaction events and sessions in their respective databases.
- Added methods for ensuring schema, upserting sessions and turns, appending events, and assigning checkpoints.
- Implemented blocking query execution and response handling for ClickHouse interactions.
- Refactored existing event handling logic to support new database backends, enhancing flexibility and performance in interaction data management.
